Resolving markdown object intersections was moved to separate class
